+++ CHANGELOG.md
@@ -36,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Consolidation of `in` as identifier in `class operator In` construct.
 - Duplicated line attribution of mid-line compiler directives.
 - Parsing of `class helper`s with parent a type.
+- Parsing of generics parameters in routine headers.
 
 ## [0.3.0] - 2024-05-29